Relocated matview models
