+++ usaspending_api/awards/models_matviews/__init__.py
@@ -1,0 +1,31 @@
+from usaspending_api.awards.models_matviews.subaward_view import SubawardView
+from usaspending_api.awards.models_matviews.summary_award_view import SummaryAwardView
+from usaspending_api.awards.models_matviews.summary_cfda_numbers_view import SummaryCfdaNumbersView
+from usaspending_api.awards.models_matviews.summary_naics_codes_view import SummaryNaicsCodesView
+from usaspending_api.awards.models_matviews.summary_psc_codes_view import SummaryPscCodesView
+from usaspending_api.awards.models_matviews.summary_state_view import SummaryStateView
+from usaspending_api.awards.models_matviews.summary_transaction_fed_acct_view import SummaryTransactionFedAcctView
+from usaspending_api.awards.models_matviews.summary_transaction_geo_view import SummaryTransactionGeoView
+from usaspending_api.awards.models_matviews.summary_transaction_month_view import SummaryTransactionMonthView
+from usaspending_api.awards.models_matviews.summary_transaction_recipient_view import SummaryTransactionRecipientView
+from usaspending_api.awards.models_matviews.summary_transaction_view import SummaryTransactionView
+from usaspending_api.awards.models_matviews.summary_view import SummaryView
+from usaspending_api.awards.models_matviews.universal_award_matview import UniversalAwardView
+from usaspending_api.awards.models_matviews.universal_transaction_matview import UniversalTransactionView
+
+__all__ = [
+    "SubawardView",
+    "SummaryAwardView",
+    "SummaryCfdaNumbersView",
+    "SummaryNaicsCodesView",
+    "SummaryPscCodesView",
+    "SummaryStateView",
+    "SummaryTransactionFedAcctView",
+    "SummaryTransactionGeoView",
+    "SummaryTransactionMonthView",
+    "SummaryTransactionRecipientView",
+    "SummaryTransactionView",
+    "SummaryView",
+    "UniversalAwardView",
+    "UniversalTransactionView",
+]
