+++ usaspending_api/awards/models_matviews/summary_transaction_geo_view.py
@@ -1,0 +1,48 @@
+from django.db import models
+
+
+class SummaryTransactionGeoView(models.Model):
+    duh = models.UUIDField(primary_key=True, help_text="Deterministic Unique Hash")
+    action_date = models.DateField()
+    fiscal_year = models.IntegerField()
+    type = models.TextField()
+    pulled_from = models.TextField()
+
+    recipient_location_country_name = models.TextField()
+    recipient_location_country_code = models.TextField()
+    recipient_location_state_code = models.TextField()
+    recipient_location_county_name = models.TextField()
+    recipient_location_county_code = models.TextField()
+    recipient_location_zip5 = models.TextField()
+    recipient_location_congressional_code = models.TextField()
+    recipient_location_city_name = models.TextField()
+
+    pop_country_name = models.TextField()
+    pop_country_code = models.TextField()
+    pop_state_code = models.TextField()
+    pop_county_name = models.TextField()
+    pop_county_code = models.TextField()
+    pop_zip5 = models.TextField()
+    pop_congressional_code = models.TextField()
+    pop_city_name = models.TextField()
+
+    awarding_agency_id = models.IntegerField()
+    funding_agency_id = models.IntegerField()
+    awarding_toptier_agency_name = models.TextField()
+    funding_toptier_agency_name = models.TextField()
+    awarding_subtier_agency_name = models.TextField()
+    funding_subtier_agency_name = models.TextField()
+    awarding_toptier_agency_abbreviation = models.TextField()
+    funding_toptier_agency_abbreviation = models.TextField()
+    awarding_subtier_agency_abbreviation = models.TextField()
+    funding_subtier_agency_abbreviation = models.TextField()
+
+    generated_pragmatic_obligation = models.DecimalField(max_digits=23, decimal_places=2, null=True, blank=True)
+    federal_action_obligation = models.DecimalField(max_digits=23, decimal_places=2, blank=True, null=True)
+    original_loan_subsidy_cost = models.DecimalField(max_digits=23, decimal_places=2, null=True, blank=True)
+    face_value_loan_guarantee = models.DecimalField(max_digits=23, decimal_places=2, null=True, blank=True)
+    counts = models.IntegerField()
+
+    class Meta:
+        managed = False
+        db_table = "summary_transaction_geo_view"
